// src/generated/models/Cr2b6_timeoffsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr2b6_timeoffscr2b6_type = {
  566210000: 'RTO',
  566210001: 'PTO',
  566210002: 'Sick'
} as const;
export type Cr2b6_timeoffscr2b6_type = keyof typeof Cr2b6_timeoffscr2b6_type;
export const Cr2b6_timeoffsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr2b6_timeoffsstatecode = keyof typeof Cr2b6_timeoffsstatecode;
export const Cr2b6_timeoffsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr2b6_timeoffsstatuscode = keyof typeof Cr2b6_timeoffsstatuscode;

export interface Cr2b6_timeoffsBase {
  "cr2b6_Employee@odata.bind"?: string;
  cr2b6_enddate?: string;
  cr2b6_name: string;
  cr2b6_note?: string;
  cr2b6_startdate?: string;
  cr2b6_timeoffid: string;
  cr2b6_type?: Cr2b6_timeoffscr2b6_type;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr2b6_timeoffsstatecode;
  statuscode?: Cr2b6_timeoffsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr2b6_timeoffs extends Cr2b6_timeoffsBase {
  cr2b6_employeename?: string;
  cr2b6_typename?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cr2b6_employee?: object;
  _cr2b6_employee_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}